+++ Changelog.md
@@ -1,9 +1,19 @@
+# 1.0.2 (Version Code: 3 - March 1st, 2026)
+
+GD-OAuth 1.0.2: Secure Authentication for OAuth Apps.
+
+## Include
+- Geode SDK v5 Support
+
+## Exclude
+- Geode SDK v4 on GD-OAuth 1.0.2
+
 # 1.0.1 (Version Code: 2 - February 24th, 2026)
 
 GD-OAuth 1.0.1: Secure Authentication for OAuth Apps.
 
-## Includes
+## Include
 - Resources
 
-## Excludes
+## Exclude
 - There is no excluded